Move util test to its own file to avoid dependencies of test_auxiliary
